ARC: cache: no need for cast in copy_to_user_page()

+++ arch/arc/include/asm/cacheflush.h
@@ -118,7 +118,7 @@ static inline int cache_is_vipt_aliasing(void)
 do {									\
 	memcpy(dst, src, len);						\
 	if (vma->vm_flags & VM_EXEC)					\
-		__sync_icache_dcache((unsigned long)(dst), vaddr, len);	\
+		__sync_icache_dcache(dst, vaddr, len);	                \
 } while (0)
 
 #define copy_from_user_page(vma, page, vaddr, dst, src, len)		\
